Update Paho lib version
