+++ Gopkg.toml
@@ -8,7 +8,7 @@
 
 [[constraint]]
   name = "github.com/eclipse/paho.mqtt.golang"
-  version = "1.1.1"
+  version = "1.2"
 
 [[constraint]]
   name = "github.com/fatih/color"
