--- tools/mqtt-bench/res-verne.txt
@@ -1,14 +1,0 @@
-========= TOTAL (1000) =========
-Total Ratio:                 1.000 (100000/100000)
-Total Runtime (sec):         33.442
-Average Runtime (sec):       30.661
-Msg time min (us):           9068.000
-Msg time max (us):           777640.000
-Msg time mean mean (us):     283473.021
-Msg time mean std (us):      11196.375
-Msg del time min (us):   0.000
-Msg del time max (us):   0.000
-Msg del time mean (us):  0.000
-Msg del time std (us):   0.000
-Average Bandwidth (msg/sec): 3.281
-Total Bandwidth (msg/sec):   3281.065
